pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-sub-generator-sakura"
version = "1.0.0"
description = "🌸 Sakura Subtitle Generator - Japanese ASR with Multi-language Translation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nelson Lin", email = "nelson.lin.tw@gmail.com"}
]
keywords = ["subtitle", "asr", "translation", "japanese", "whisper", "ml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License", 
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # Core ML libraries
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.0.0",
    # Audio processing
    "librosa>=0.10.0",
    "ffmpeg-python>=0.2.0",
    # ASR and NLP
    "accelerate>=0.20.0",
    "sentencepiece>=0.1.99",
    "sacremoses>=0.0.53",
    # Utilities
    "numpy>=1.21.0",
    "tqdm>=4.65.0",
    # Chinese text processing
    "opencc-python-reimplemented>=0.1.7",
    # SakuraLLM GGUF support
    "llama-cpp-python>=0.2.0",
    # Optional optimizations
    "optimum[onnxruntime]>=1.8.0",
]

[project.optional-dependencies]
# Development tools are managed via project-scoped `uv tool run` - see Makefile
# This keeps the project environment clean without global installation
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pre-commit>=3.0.0",
]

gpu = [
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
]

# Apple Silicon optimized dependencies
apple-silicon = [
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0", 
    "torchaudio>=2.0.0",
    # Metal Performance Shaders optimizations
    "accelerate>=0.20.0",
    # Optimized for ARM64 architecture
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    # Fast tokenizers with ARM64 support
    "tokenizers>=0.13.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
]

all = [
    "py-sub-generator-sakura[dev,gpu,test]"
]

all-apple = [
    "py-sub-generator-sakura[dev,apple-silicon,test]"
]

[project.urls]
Homepage = "https://github.com/yc0/py-sub-generator-sakura"
Repository = "https://github.com/yc0/py-sub-generator-sakura.git"
"Bug Tracker" = "https://github.com/yc0/py-sub-generator-sakura/issues"

[project.scripts]
sakura-subtitle = "main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# Modern dependency groups (development tools managed via uv tool)
[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "pytest-mock>=3.10.0",
    "pre-commit>=3.0.0",
]

# UV-specific configuration
[tool.uv]
# Development tools (ruff, black, isort, mypy) are managed via project-scoped `uv tool run`
# This approach keeps them separate from project dependencies while avoiding global installation.
# Tools are executed on-demand via: uv tool run <tool>@latest
# Use: make lint, make format, make typecheck

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I", "N", "UP"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--cov=src",
    "--cov-report=html",
    "--cov-report=term-missing",
]

# Hatchling build configuration
[tool.hatch.build.targets.wheel]
packages = ["src"]
include = [
    "main.py",
    "config.json",
    "src/**/*.py"
]

[tool.hatch.build.targets.sdist]
include = [
    "main.py",
    "config.json", 
    "src/",
    "README.md",
    "pyproject.toml"
]
